make font selection more robust
deleting the buffer being used to store font data wasn’t being properly handled - the font defintion we set up and store in the `fonts` list would be removed, but the font would remain set on the canvas and in use.  once the psram for the buffer storing the font data was re-assigned all bets would be off…

these changes fix that - the underlying data is now tracked along with the font object inside the context.  as a result fonts currently in use will remain available, even if they’ve been deleted out of the `fonts` list.  smart pointers are used for storing font objects inside the context (same as in the `fonts` list), so they will stick around for as long as we need them.